move test cases to collection.rs
